% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFAULT_control1.R
\docType{data}
\name{DEFAULT_control1}
\alias{DEFAULT_control1}
\title{A Default Value for corp_or_dtm 1}
\format{An object of class \code{list} of length 2.}
\usage{
DEFAULT_control1
}
\description{
This list object is by default called by \code{\link{corp_or_dtm}}.
}
\details{
The object specifies word length from 1 to 
25, and takes a regular expression as tokenizer. Also, \code{DEFAULT_control2}
sets lengh from 2 to 25.
}
\examples{
require(tm)
x <- c(
  "Hello, what do you want to drink?", 
  "drink a bottle of milk", 
  "drink a cup of coffee", 
  "drink some water")
dtm <- corp_or_dtm(x, from = "v", type = "dtm")
dtm <- corp_or_dtm(x, from = "v", type = "dtm", control = DEFAULT_control1)
another <- DEFAULT_control1
another$wordLengths <- c(4, 20)
dtm <- corp_or_dtm(x, from = "v", type = "dtm", control = another)
}
\keyword{datasets}
