% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEFAULT_cutter.R
\docType{data}
\name{DEFAULT_cutter}
\alias{DEFAULT_cutter}
\title{A Default Cutter}
\format{
An object of class \code{jiebar} (inherits from \code{segment}, \code{jieba}) of length 11.
}
\usage{
DEFAULT_cutter
}
\description{
This is simply a jiebar object created when the package is loaded. 
\code{write} is set to \code{FALSE}, so as to prevent segmented text from 
being automatically written into disk.
}
\examples{
require(jiebaR)
x <- c("drink a bottle of milk", 
  "drink a cup of coffee", 
 "drink some water")
seg_file(x, from = "v")
seg_file(x, from = "v", mycutter = DEFAULT_cutter)
}
\keyword{datasets}
